[CC-31] Use English text as message keys for licenses
The legalcode translation files have been updated to use the
English text of the messages as the keys for the other
languages' translations.

The directory structure has also been modified so that
Django will be able to find these files if we add
``legalcode`` to the locale paths.  (Django insists
on the ``LC_MESSAGES`` directory in every language
directory. I don't know what use it is, but we need
it.)